+++ occurrence-processor/src/main/java/org/gbif/occurrence/processor/InterpretedProcessor.java
@@ -131,8 +131,8 @@ public class InterpretedProcessor {
    * @param attemptId     the crawl attempt id, only used for passing along in logs and subsequent messages.
    * @param datasetKey    the dataset that this occurrence belongs to (must not be null if fromCrawl is true)
    */
-  public void buildInterpreted(int occurrenceKey, OccurrencePersistenceStatus status, boolean fromCrawl,
-    @Nullable Integer attemptId, @Nullable UUID datasetKey) {
+  public void buildInterpreted(Long occurrenceKey, OccurrencePersistenceStatus status, boolean fromCrawl,
+                               @Nullable Integer attemptId, @Nullable UUID datasetKey) {
     checkArgument(occurrenceKey > 0, "occurrenceKey must be greater than 0");
     checkNotNull(status, "status can't be null");
     if (fromCrawl) {
@@ -260,7 +260,7 @@ public class InterpretedProcessor {
     return verbatimOccurrence;
   }
 
-  private void logError(String message, int occurrenceKey, UUID datasetKey, boolean fromCrawl) {
+  private void logError(String message, Long occurrenceKey, UUID datasetKey, boolean fromCrawl) {
     // TODO: send msg?
     LOG.error("{} verbatim occurrence with key [{}] - skipping.", message, occurrenceKey);
     if (fromCrawl) {
